Convert mime data length to use unsigned long
What we want to use is size_t, but we don't want the implied POSIX
dependency. However, POSIX does say that size_t is an unsigned integer
that is no longer than a long, so it would appear safe to use an
unsigned long as a replacement. Safer at least than unsigned int.

+++ src/cairo-output-stream-private.h
@@ -175,7 +175,7 @@ _cairo_memory_stream_length (cairo_output_stream_t *stream);
 cairo_private cairo_status_t
 _cairo_memory_stream_destroy (cairo_output_stream_t *abstract_stream,
 			      unsigned char **data_out,
-			      unsigned int *length_out);
+			      unsigned long *length_out);
 
 cairo_private cairo_output_stream_t *
 _cairo_null_stream_create (void);
